Corrected UpdateGrade to only allow 1-5 ints

+++ app/Contracts/Services/InspectedComponentServiceContract.php
@@ -6,5 +6,5 @@ namespace App\Contracts\Services;
 
 interface InspectedComponentServiceContract
 {
-    public function updateGrade(string $id, float $grade): void;
+    public function updateGrade(string $id, int $grade): void;
 }
